Удалены ненужные case из главного файла index.ts, реализован переход в Чаты в компоненте BackInChats, в Чатах реализован переход в Профиль
