--- src/components/BackIiChats/backInChats.hbs
@@ -1,10 +1,0 @@
-<div class="profile-back-page">
-    <a href="{{{route}}}">
-        <div class="arrow">
-            <svg width="7" height="12" viewBox="0 0 7 12" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path d="M6 11L2 6L6 1" stroke="white" stroke-width="1.6"></path>
-            </svg>
-        </div>
-
-    </a>
-</div>
