+++ src/components/BackInChats/backInChats.hbs
@@ -1,0 +1,3 @@
+<div class="profile-back-page">
+    {{{backInChatsButton}}}
+</div>
